Improve MMIO support to utilize the HAL framework.
