+++ configure.ac
@@ -203,7 +203,7 @@ then
     AM_CFLAGS="$AM_CFLAGS -DWOLFTPM_I2C"
 fi
 
-# WOLFTPM_MMIO_BUILTIN_CB
+# Memory-mapped I/O
 AC_ARG_ENABLE([mmio],
     [AS_HELP_STRING([--enable-mmio],[Enable built-in MMIO callbacks (default: disabled)])],
     [ ENABLED_MMIO=$enableval ],
@@ -212,7 +212,7 @@ AC_ARG_ENABLE([mmio],
 
 if test "x$ENABLED_MMIO" = "xyes"
 then
-    AM_CFLAGS="$AM_CFLAGS -DWOLFTPM_MMIO_BUILTIN_CB"
+    AM_CFLAGS="$AM_CFLAGS -DWOLFTPM_MMIO"
 fi
 
 # Advanced IO
@@ -441,7 +441,7 @@ AM_CONDITIONAL([BUILD_WINAPI], [test "x$ENABLED_WINAPI" = "xyes"])
 AM_CONDITIONAL([BUILD_NUVOTON], [test "x$ENABLED_NUVOTON" = "xyes"])
 AM_CONDITIONAL([BUILD_CHECKWAITSTATE], [test "x$ENABLED_CHECKWAITSTATE" = "xyes"])
 AM_CONDITIONAL([BUILD_AUTODETECT], [test "x$ENABLED_AUTODETECT" = "xyes"])
-AM_CONDITIONAL([BUILD_HAL], [test "x$ENABLED_EXAMPLE_HAL" = "xyes"])
+AM_CONDITIONAL([BUILD_HAL], [test "x$ENABLED_EXAMPLE_HAL" = "xyes" || test "x$ENABLED_MMIO" = "xyes"])
 
 
 CREATE_HEX_VERSION
