+++ hal/include.am
@@ -8,6 +8,7 @@ src_libwolftpm_la_SOURCES += \
                     hal/tpm_io_barebox.c \
                     hal/tpm_io_linux.c \
                     hal/tpm_io_infineon.c \
+                    hal/tpm_io_mmio.c \
                     hal/tpm_io_microchip.c \
                     hal/tpm_io_st.c \
                     hal/tpm_io_qnx.c \
